+++ hatanaka/__init__.py
@@ -2,4 +2,4 @@ from .general_compression import *
 from .hatanaka import *
 
 __version__ = '2.6.0'
-rnxcmp_version = '4.0.8'
+rnxcmp_version = '4.1.0'